--- src/main/java/com/example/dp/domain/user/dto/UserCheckCodeRequestDto.java
@@ -1,13 +1,0 @@
-package com.example.dp.domain.user.dto;
-
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class UserCheckCodeRequestDto {
-
-    private String email;
-    private String code;
-}
